+++ ignite/cmd/plugin.go
@@ -10,7 +10,10 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
+	pluginsconfig "github.com/ignite/cli/ignite/config/plugins"
+	"github.com/ignite/cli/ignite/pkg/clictx"
 	"github.com/ignite/cli/ignite/pkg/cliui"
+	"github.com/ignite/cli/ignite/pkg/cliui/icons"
 	"github.com/ignite/cli/ignite/pkg/xgit"
 	"github.com/ignite/cli/ignite/services/plugin"
 )
@@ -26,17 +29,33 @@ const (
 // LoadPlugins tries to load all the plugins found in configuration.
 // If no configuration found, it returns w/o error.
 func LoadPlugins(ctx context.Context, rootCmd *cobra.Command) error {
-	// NOTE(tb) Not sure if it's the right place to load this.
-	chain, err := newChainWithHomeFlags(rootCmd)
+	cfg, _, err := parseLocalPlugins(rootCmd)
 	if err != nil {
-		// Binary is run outside of an chain app, plugins can't be loaded
+		// if binary is run where there is no plugins.yml, don't load
 		return nil
 	}
-	plugins, err = plugin.Load(ctx, chain)
+
+	// TODO: parse global config
+
+	plugins, err = plugin.Load(ctx, cfg)
 	if err != nil {
 		return err
 	}
 	return loadPlugins(rootCmd, plugins)
+}
+
+func parseLocalPlugins(rootCmd *cobra.Command) (cfg *pluginsconfig.Config, path string, err error) {
+	appPath := flagGetPath(rootCmd)
+	pluginsPath := getPlugins(rootCmd)
+	if pluginsPath == "" {
+		if pluginsPath, err = pluginsconfig.LocateDefault(appPath); err != nil {
+			return cfg, appPath, err
+		}
+	}
+
+	cfg, err = pluginsconfig.ParseFile(pluginsPath)
+
+	return cfg, pluginsPath, err
 }
 
 func loadPlugins(rootCmd *cobra.Command, plugins []*plugin.Plugin) error {
@@ -241,23 +260,22 @@ func linkPluginCmd(rootCmd *cobra.Command, p *plugin.Plugin, pluginCmd plugin.Co
 	if len(pluginCmd.Commands) == 0 {
 		// pluginCmd has no sub commands, so it's runnable
 		newCmd.RunE = func(cmd *cobra.Command, args []string) error {
-			execCmd := plugin.ExecutedCommand{
-				Use:  cmd.Use,
-				Path: cmd.CommandPath(),
-				Args: args,
-				With: p.With,
-			}
-			execCmd.SetFlags(cmd.Flags())
-			// Call the plugin Execute
-			err := p.Interface.Execute(execCmd)
-			// NOTE(tb): This pause gives enough time for go-plugin to sync the
-			// output from stdout/stderr of the plugin. Without that pause, this
-			// output can be discarded and not printed in the user console.
-			time.Sleep(100 * time.Millisecond)
-			if err != nil {
-				return fmt.Errorf("plugin %q Execute() error : %w", p.Path, err)
-			}
-			return nil
+			return clictx.Do(cmd.Context(), func() error {
+				execCmd := plugin.ExecutedCommand{
+					Use:  cmd.Use,
+					Path: cmd.CommandPath(),
+					Args: args,
+					With: p.With,
+				}
+				execCmd.SetFlags(cmd.Flags())
+				// Call the plugin Execute
+				err := p.Interface.Execute(execCmd)
+				// NOTE(tb): This pause gives enough time for go-plugin to sync the
+				// output from stdout/stderr of the plugin. Without that pause, this
+				// output can be discarded and not printed in the user console.
+				time.Sleep(100 * time.Millisecond)
+				return err
+			})
 		}
 	} else {
 		for _, pluginCmd := range pluginCmd.Commands {
@@ -293,6 +311,9 @@ func NewPlugin() *cobra.Command {
 	c.AddCommand(NewPluginUpdate())
 	c.AddCommand(NewPluginScaffold())
 	c.AddCommand(NewPluginDescribe())
+	c.AddCommand(NewPluginAdd())
+	c.AddCommand(NewPluginRemove())
+
 	return c
 }
 
@@ -341,6 +362,118 @@ func NewPluginUpdate() *cobra.Command {
 			return errors.Errorf("Plugin %q not found", args[0])
 		},
 	}
+}
+
+func NewPluginAdd() *cobra.Command {
+	cmdPluginAdd := &cobra.Command{
+		Use:   "add [path] [args]",
+		Short: "Adds a plugin declaration to a chain's plugin configuration",
+		Long: `Adds a plugin declaration to a chain's plugin configuration.
+				Respects key value pairs declared after the plugin path to be added to the generated configurationdefinition
+				example:
+					ignite plugin add /path/to/plugin/ foo:bar
+		`,
+		Args: cobra.MinimumNArgs(1),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			s := cliui.New(cliui.WithStdout(os.Stdout))
+
+			conf, persistPath, err := parseLocalPlugins(cmd)
+			if err != nil {
+				return err
+			}
+
+			for _, p := range conf.Plugins {
+				if p.Path == args[0] {
+					return fmt.Errorf("cannot add duplicate plugin %s", args[0])
+				}
+			}
+
+			p := pluginsconfig.Plugin{
+				Path: args[0],
+				With: make(map[string]string),
+			}
+
+			var pluginArgs []string
+			if len(args) > 1 {
+				pluginArgs = args[1:]
+			}
+
+			for _, pa := range pluginArgs {
+				kv := strings.Split(pa, ":")
+				if len(kv) != 2 {
+					continue
+				}
+				p.With[kv[0]] = kv[1]
+			}
+
+			ctx := context.Background()
+			s.StartSpinner("Loading plugin")
+			pluginInstance, err := plugin.LoadSingle(ctx, &p)
+			s.StopSpinner()
+
+			if err != nil {
+				return err
+			}
+
+			if pluginInstance.Error != nil {
+				return errors.Wrapf(pluginInstance.Error, fmt.Sprintf("Error while attempting to load plugin: %s\n", args[0]))
+			}
+			s.Println("Done loading plugin")
+
+			conf.Plugins = append(conf.Plugins, p)
+
+			if err != nil {
+				return err
+			}
+
+			err = conf.Save(persistPath)
+
+			if err != nil {
+				return err
+			}
+
+			s.Printf("🎉 %s added \n", args[0])
+
+			return nil
+		},
+	}
+
+	return cmdPluginAdd
+}
+
+func NewPluginRemove() *cobra.Command {
+	cmdPluginRemove := &cobra.Command{
+		Use:   "remove [path]",
+		Short: "Removes a plugin declaration from a chain's plugin configuration",
+		Args:  cobra.ExactArgs(1),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			s := cliui.New(cliui.WithStdout(os.Stdout))
+
+			conf, persistPath, err := parseLocalPlugins(cmd)
+			if err != nil {
+				return err
+			}
+
+			for i, cp := range conf.Plugins {
+				if cp.Path == args[0] {
+					conf.Plugins = append(conf.Plugins[:i], conf.Plugins[i+1:]...)
+					break
+				}
+			}
+
+			err = conf.Save(persistPath)
+
+			if err != nil {
+				return err
+			}
+
+			s.Printf("%s %s removed\n", icons.OK, args[0])
+
+			return nil
+		},
+	}
+
+	return cmdPluginRemove
 }
 
 func NewPluginScaffold() *cobra.Command {