+++ ignite/pkg/cliui/colors/colors.go
@@ -12,39 +12,60 @@ const (
 	Green   = "#4e9a06"
 	Magenta = "#75507b"
 	Cyan    = "#34e2e2"
+	White   = "#FFFFFF"
+	HiBlue  = "#729FCF"
 )
 
 var (
-	info     = lipgloss.NewStyle().Foreground(lipgloss.Color(Yellow)).Render
-	infof    = lipgloss.NewStyle().Foreground(lipgloss.Color(Yellow)).Render
-	error    = lipgloss.NewStyle().Foreground(lipgloss.Color(Red)).Render
-	success  = lipgloss.NewStyle().Foreground(lipgloss.Color(Green)).Render
-	modified = lipgloss.NewStyle().Foreground(lipgloss.Color(Magenta)).Render
+	info     = lipgloss.NewStyle().Foreground(lipgloss.Color(Yellow))
+	infof    = lipgloss.NewStyle().Foreground(lipgloss.Color(Yellow))
+	error    = lipgloss.NewStyle().Foreground(lipgloss.Color(Red))
+	success  = lipgloss.NewStyle().Foreground(lipgloss.Color(Green))
+	modified = lipgloss.NewStyle().Foreground(lipgloss.Color(Magenta))
+	name     = lipgloss.NewStyle().Foreground(lipgloss.Color(White)).Bold(true)
+	mnemonic = lipgloss.NewStyle().Foreground(lipgloss.Color(HiBlue))
+	faint    = lipgloss.NewStyle().Faint(true)
 )
 
-func SprintFunc(code string) func(i ...interface{}) string {
+// SprintFunc returns a function to apply a foreground color to any number of texts.
+// The retuned function receives strings as arguments with the text that should be colorized.
+// Color specifies a color by hex or ANSI value.
+func SprintFunc(color string) func(i ...interface{}) string {
 	return func(i ...interface{}) string {
-		render := lipgloss.NewStyle().Foreground(lipgloss.Color(code)).Render
-		return render(fmt.Sprint(i...))
+		style := lipgloss.NewStyle().Foreground(lipgloss.Color(color))
+		return style.Render(fmt.Sprint(i...))
 	}
 }
 
 func Info(i ...interface{}) string {
-	return info(fmt.Sprint(i...))
+	return info.Render(fmt.Sprint(i...))
 }
 
-func Infof(i ...interface{}) string {
-	return infof(fmt.Sprint(i...))
+func Infof(format string, i ...interface{}) string {
+	return infof.Render(fmt.Sprintf(format, i...))
 }
 
 func Error(i ...interface{}) string {
-	return error(fmt.Sprint(i...))
+	return error.Render(fmt.Sprint(i...))
 }
 
 func Success(i ...interface{}) string {
-	return success(fmt.Sprint(i...))
+	return success.Render(fmt.Sprint(i...))
 }
 
 func Modified(i ...interface{}) string {
-	return modified(fmt.Sprint(i...))
+	return modified.Render(fmt.Sprint(i...))
+}
+
+func Name(i ...interface{}) string {
+	return name.Render(fmt.Sprint(i...))
+}
+
+func Mnemonic(i ...interface{}) string {
+	return mnemonic.Render(fmt.Sprint(i...))
+}
+
+// Faint styles the text using a dimmer shade for the foreground color.
+func Faint(i ...interface{}) string {
+	return faint.Render(fmt.Sprint(i...))
 }