+++ ignite/templates/field/datatype/bool.go
@@ -3,7 +3,10 @@ package datatype
 import (
 	"fmt"
 
+	"github.com/emicklei/proto"
+
 	"github.com/ignite/cli/ignite/pkg/multiformatname"
+	"github.com/ignite/cli/ignite/pkg/protoanalysis/protoutil"
 )
 
 // DataBool bool data type definition
@@ -34,6 +37,9 @@ var DataBool = DataType{
 	},
 	ToString: func(name string) string {
 		return fmt.Sprintf("strconv.FormatBool(%s)", name)
+	},
+	ToProtoField: func(_, name string, index int) *proto.NormalField {
+		return protoutil.NewField(name, "bool", index)
 	},
 	GoCLIImports: []GoImport{{Name: "github.com/spf13/cast"}},
 }